Fix 1.2 migration on postgresql.
Django supposedly already wraps RunPython in an atomic transaction, and
it won't let you call commit or autocommit from within an atomic
transaction (I think the entire migration is also wrapped in a
transaction).

This is my hack to commit the schema alterations prior to the data
changes.

I've only tested this on postgresql, it probably needs tested on mysql
and sqlite prior to merging.

Kegbot/kegbot-server#325